学生课程维护、重修退课成绩校验添加（master 2019-09-26） 黎光明--001

See merge request tongji/course-election!1548

+++ src/main/java/com/server/edu/election/service/impl/RetakeCourseServiceImpl.java
@@ -271,6 +271,12 @@ public class RetakeCourseServiceImpl implements RetakeCourseService {
             Long id = rebuildCourseVo.getTeachingClassId();
             List<Long> list = new ArrayList<>(1);
             list.add(id);
+            List<String> courseCodes = new ArrayList<>(1);
+            courseCodes.add(courseCode);
+            List<String> courses = ScoreServiceInvoker.findCourseHaveScore(studentId, calendarId, courseCodes);
+            if (CollectionUtil.isNotEmpty(courses)) {
+                throw new ParameterValidateException(I18nUtil.getMsg(I18nUtil.getMsg("elcCourseUphold.removeCourseError",courses.get(0))));
+            }
             courseTakeDao.deleteCourseTask(list, studentId);
             teachingClassDao.decrElcNumber(teachingClassId);
             // 添加选课日志
